Done with refactoring for now
