+++ include/tests.hpp
@@ -1,0 +1,19 @@
+/**
+ * @file tests.hpp
+ * @author Hassan Sarhan (hassanAsarhan@outlook.com)
+ * @brief This file defines various tests
+ * @date 2023-07-20
+ *
+ * @copyright Copyright (c) 2023
+ *
+ */
+
+#ifndef TESTS_HPP
+#define TESTS_HPP
+
+/**
+ * @brief Tests for the different validation functions
+ */
+void inputValidatorTests();
+
+#endif
